backend/requirements.txt
fastapi
uvicorn
roboflow
opencv-python
scikit-learn
numpy
python-dotenv